+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.36.0] - 2024-12-23
+### Added
+- Meeting class
+- MeetingJoinParameters class
+- 'meetings' property in :obj:models.responses.CalendarResponse
+
 ## [2.35.0] - 2024-11-29
 ### Added
 - 'person_id' parameter to PyrusAPI client for auth method